add a check for total travel time
inside generating solver space travel matrix, add a check for total
trip travel time is less than horizon, otherwise trip cannot be
performed.

The travel time matrix must therefore be the final arbiter of how big
the problem is, as the list of nodes in the demand object will be too
long by the set of infeasible pairs.