conditional constraint seems to do the job
Added a count dimension to flag early visits when breaks need to be
on.  Then used "ActiveVar()" type constraint to make sure a node is
active and being served, that the count is >0 and <3 in order to say
break must be on.  The key thing I forgot is that you must multiply
dimensions by "ActiveVar()" in order to make sure that the value
really is real, otherwise the solver can float along a value in the
register that is meaningless.  Quirk I forgot about.