more changes lumped together
lots of minor changes everywhere to set up multiple break nodes, if
needed.  Basically any link greater than or equal to 660 minutes can
need two or more breaks, so now that can happen.  Each successive
break node get progressively "closer" to the destination, but it is
all dummy values and not real.  No relation to the location of the
real break.

Conditional constraints to put the dummy breaks into use.