rewrite initial routes code
using stored sequence of breaks now, rather than figuring it out from
the travel time matrix.  Although that worked, it would be impossible
to debug.

Note that this new way is still pretty confusing and will be difficult
to debug, but at least it is possible...the old way was not.