maybe?
hackity hack.  basically, in an 11 hr period, you'll hit both an 8hr,
30 min break, and an 11 hr, 10 hr break.  Don't reset the short break
tracker by 8hrs for both.  rather, if two in a row, reset the first
with 3 hrs, and the second with 8, totaling 11.

I have to do first 3, second 11 because I only know the nodes from and
to, and I'm tracking break time as a service time at the from node,
and adding it after leaving, instead of before arriving.  I could
rewrite everything to fix that, but these are dummy nodes anyway.